Simplified data model, changed productNames on shipments/packaging to be nondynamic
Placed a column for userId directly on packaging/shipments to allow for cascades if a user account is removed. Also removes the need for resource intensive table joins. Changed the GET/POST for shipments/packaging to acquire productNames as a nondynamic string from the time package is added. Shipments retrieves this from packaging when the shipment is added.