Added data seeding of productUUIDs and names on shipments. Added getdetails products route.
All shipments/packages are linked to products in the productPackages table. This relationships table is used to add strings for each name and UUID related to each shipment. A getdetails route on products will give information about all found shipments for any given product as well as the product data.